Add support for EPM publication
This commit modifies truffle.js to detect a secrets.json and, if one
exists, instantiate an hdwallet provider capable of deploying an EPM
package to the Ropsten testnet. The secrets.json file should look like:

{
  "mnemonic": "igor cannot dunk ..."
}

Important to note, there is an upstream bug in Truffle at the time of
this commit which causes EPM deployment to fail when it detects multiple
networks specified in the truffle.js file. Until this is fixed, to
deploy to EPM you will need to comment out all the networks other than Ropsten
before running `truffle publish`.