Modify tests to check for EVM errors on failure
This is not the fancy Simon-approved way of checking for throws.
However. In a few short weeks the reverts we have in place will return
real error messages to Truffle and then we'll be able to make small
changes to this script such that checking for reversion in this way
won't be at all hacky, but in fact proper.

I don't want to go through the complexity of testing exceptions in the
more rigorous way when Metro will be out so soon... And really this
is only slightly hacky.